Refactor CLI structure: split commands and utilities into separate files
- Moved individual commands into their own files for better modularity
- Extracted common utilities into a separate package for reuse across commands
- Reduced the size of the main CLI file for improved readability and maintainability
- Simplified the process of adding new commands by creating dedicated files for each
- Ensured all tests pass and logic remains intact
