+++ cmd/delete.go
@@ -1,0 +1,48 @@
+package cmd
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/ollama/ollama/api"
+	"github.com/spf13/cobra"
+)
+
+func NewDeleteCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "rm MODEL [MODEL...]",
+		Short:   "Remove a model",
+		Args:    cobra.MinimumNArgs(1),
+		PreRunE: checkServerHeartbeat,
+		RunE:    deleteHandler,
+	}
+
+	return cmd
+}
+
+func deleteHandler(cmd *cobra.Command, args []string) error {
+	client, err := api.ClientFromEnvironment()
+	if err != nil {
+		return err
+	}
+
+	// Unload the model if it's running before deletion
+	opts := &runOptions{
+		Model:     args[0],
+		KeepAlive: &api.Duration{Duration: 0},
+	}
+	if err := loadOrUnloadModel(cmd, opts); err != nil {
+		if !strings.Contains(err.Error(), "not found") {
+			return fmt.Errorf("unable to stop existing running model \"%s\": %s", args[0], err)
+		}
+	}
+
+	for _, name := range args {
+		req := api.DeleteRequest{Name: name}
+		if err := client.Delete(cmd.Context(), &req); err != nil {
+			return err
+		}
+		fmt.Printf("deleted '%s'\n", name)
+	}
+	return nil
+}
